task_wdt: compute the fallback window as an unsigned value
wdt_timeout_cfg.window.max is a uint32_t, but both Kconfig symbols
feeding it are ints. The sum is therefore evaluated in int and
overflows before the result is converted, which is undefined behaviour
as soon as the two timeouts add up to more than INT_MAX milliseconds.

Cast both operands so that the addition is carried out in uint32_t,
where the largest sum the two symbols can produce still fits.

+++ subsys/task_wdt/task_wdt.c
@@ -134,8 +134,8 @@ int task_wdt_init(const struct device *hw_wdt)
 
 		wdt_config.flags = WDT_FLAG_RESET_SOC;
 		wdt_config.window.min = 0U;
-		wdt_config.window.max = CONFIG_TASK_WDT_MIN_TIMEOUT +
-			CONFIG_TASK_WDT_HW_FALLBACK_DELAY;
+		wdt_config.window.max = (uint32_t)CONFIG_TASK_WDT_MIN_TIMEOUT +
+					(uint32_t)CONFIG_TASK_WDT_HW_FALLBACK_DELAY;
 		wdt_config.callback = NULL;
 
 		hw_wdt_dev = hw_wdt;
